adding validations to the User controller fields

+++ src/main/java/com/ordering/controllers/UserController.java
@@ -2,19 +2,26 @@ package com.ordering.controllers;
 
 import com.ordering.dto.UserDto;
 import com.ordering.services.UserService;
+import com.ordering.validations.ErrorValidationComponent;
+import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+
 @Slf4j
 @RestController
 @RequestMapping("/api/user")
 public class UserController {
-private UserService userService;
+    UserService userService;
+    ErrorValidationComponent errorValidationComponent;
+
     @Autowired
-    public UserController(UserService userService){
-        this.userService=userService;
+    public UserController(UserService userService, ErrorValidationComponent errorValidationComponent) {
+        this.userService = userService;
+        this.errorValidationComponent = errorValidationComponent;
     }
 
     @GetMapping()
@@ -22,25 +29,33 @@ private UserService userService;
         log.info("Search for all Users");
         return new ResponseEntity<>(userService.getAllUsers(), HttpStatus.OK);
     }
+
     @GetMapping("/{email}")
-    public ResponseEntity<?> getUserByEmail(@PathVariable String email) {
-        log.info("User Search by Email: "+ email);
+    public ResponseEntity<?> getUserByEmail(@Valid @PathVariable String email, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) return errorValidationComponent.validationErrors(bindingResult);
+        log.info("User Search by Email: " + email);
         return new ResponseEntity<>(userService.getUserByEmail(email), HttpStatus.OK);
     }
+
     @PostMapping("")
-    public ResponseEntity<?> createUser(@RequestBody UserDto user) {
-        log.info("Creating user: "+ user.getName());
-        return new ResponseEntity<>(userService.createUser(user),HttpStatus.CREATED);
+    public ResponseEntity<?> createUser(@Valid @RequestBody UserDto user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) return errorValidationComponent.validationErrors(bindingResult);
+        log.info("Creating user: " + user.getName());
+        return new ResponseEntity<>(userService.createUser(user), HttpStatus.CREATED);
     }
+
     @PutMapping("")
-    public ResponseEntity<?> updateUser(@RequestBody UserDto user) {
-        log.info("Updating user: "+ user.getName());
-        return new ResponseEntity<>(userService.updateUser(user),HttpStatus.OK);
+    public ResponseEntity<?> updateUser(@Valid @RequestBody UserDto user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) return errorValidationComponent.validationErrors(bindingResult);
+        log.info("Updating user: " + user.getName());
+        return new ResponseEntity<>(userService.updateUser(user), HttpStatus.OK);
     }
+
     @DeleteMapping("/{email}")
-    public ResponseEntity<?> deleteUser(@PathVariable String email) {
-        log.info("Deleting user with email: "+ email);
-        return new ResponseEntity<>(userService.deleteUserByEmail(email),HttpStatus.OK);
+    public ResponseEntity<?> deleteUser(@Valid @PathVariable String email,BindingResult bindingResult) {
+        if (bindingResult.hasErrors())return errorValidationComponent.validationErrors(bindingResult);
+        log.info("Deleting user with email: " + email);
+        return new ResponseEntity<>(userService.deleteUserByEmail(email), HttpStatus.OK);
     }
 
 }
